pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "strands-agent-tui"
version = "0.1.0"
description = "Interactive TUI coding-agent platform prototype built on the Strands Agent SDK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "strands-agents[openai]",
  "textual>=0.61.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
]

[project.scripts]
strands-agent = "strands_agent_tui.app:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
